refactor(UnsnoozeTripsJob): Extract UnsnoozeTripsJob.
